requirements.txt
# Core
Flask==3.0.3
Werkzeug==3.0.3

# Auth
Flask-Login==0.6.3

# Security
Flask-WTF==1.2.1          # CSRF protection
Flask-Limiter==3.7.0      # Rate limiting

# Crypto
cryptography==42.0.8

# Scheduling
schedule==1.2.2

# Data / conversion
pandas==2.2.2
openpyxl==3.1.4
PyYAML==6.0.1

# HTTP
requests==2.32.3
urllib3==2.2.2

# Config
python-dotenv==1.0.1

# Production WSGI server
gunicorn==22.0.0
